Fix no_log bug with result lists
The combination of with_items, register, and no_log meant that we
were modifying a list which was a shared reference between the
results object used by ansible and that which we log to json.
Make a deep copy of the results object before we modify it so that
we don't modify the "original".

Also, correct a comment about the location of an import.

This adds a test which fails without the fix.

Change-Id: Iaab94f4ac8a0f58089912e464f6dfcf2e5f8ce71