Report to all reporters even if one fails
We short circuited reporting if any one reported failed. Instead do our
best to report to all reporters even if any one fails this way the
message has a better chance of getting out. This issue was identified
during the openstack zuul v3 transition as an item that needed to be
fixed.

Change-Id: Ia7f14fb108cd9ecdb14f6cada29abdf69acc1ed2
Story: 2001441