Update mysql fixture for modern mysql
There are two changes we are making to accomodate more modern mysql. The
first is apparently it is now default on newer mysqls to not do dns
lookups by default. This means we should use 127.0.0.1 instead of
localhost as the connection address as a dns lookup on localhost will
fail by default on some mysqls.

The other change we make is to allow the openstack_citest user access
from any address because even if talking to localhost/127.0.0.1 the
database may not live there (as in the case of a docker container with a
port forward, the container sees your docker network address as source
of connection not localhost).

Change-Id: I8638fea4c7ee89a4f41decffd595081c0f9bd888