Support regex matching of github status
The github status requirements matching and trigger filter are
currently plain text matching based. This currently limits sharing of
pipeline definitions between tenants as zuul reports the status as
'<tenant>/<pipeline>'. This currently makes it necessary to define
trigger filter for each tenant [1] and completely blocks pipeline
requirements.

A solution to this is regex matching which makes it possible to define
the filter once [2].

Further this enables an interesting further use case to trigger on any
successfull status [3]. This makes it easier to cooperate with other
CI systems or github apps which also set a status.

Directly use re2 as this will be used in the future for regex
matching.

[1] Trigger filter snippet
  trigger:
    github:
      - event: pull_request
        action: status
        status:
          - zuul:tenant1/check:success
          - zuul:tenant2/check:success
          - zuul:tenant3/check:success
          - zuul:tenant4/check:success

[2] Regex trigger filter snippet
  trigger:
    github:
      - event: pull_request
        action: status
        status:
          - zuul:.+/check:success

[3] Generic success filter snippet
  trigger:
    github:
      - event: pull_request
        action: status
        status:
          - .*:success

Change-Id: Id1b9d7334db78d0f13db33d47a80ffdb65f921df