Remove extra words from gerrit setup docs
While the word "review" is sometimes casually used to describe a
change in Gerrit, the correct noun is "change" (which this sentence
already used).  A "review" is something that someone (or Zuul) later
does to the change.  Let's not confuse them in docs (which could be
misleading to a new user).

Change-Id: I50510c4f2650cf30217e1f182dc289cc4f3e4682