Coerce MappingProxyTypes in job vars to dicts
Mappingproxies are not json serializable, so go ahead and convert
them to dicts when we are freezing the job.  Some of them need
already need to be converted in order to override values, so we
should be consistent in the output.

Also, because we were testing only that it was a dict instance,
the update wouldn't have worked anyway.  Fix that, and expand a
test to cover deep dictionaries as job vars.

Change-Id: Ie34b8b583b0bdb2bcab4bac4d6b45b474cfcc8fe