Add additional steps for configuring Zuul services on CentOS 7
This change adds the additional steps to follow while installing Zuul on
CentOS 7. The included steps are
- creation of zuul-scheduler.service.d, zuul-executor.service.d, and
zuul-web.service.d folders in /etc/systemd/system to place provided
drop-in files
- copying provided drop-in files to appropriate locations so ExecStart
and ExecStop commands can be overriden by systemd

Apart from the additional steps, the package installation command on Fedora27
is switched to yum to keep the differences between Fedora27 and CentOS 7 at
minimum.

Depends-On: https://review.openstack.org/#/c/565074/
Change-Id: I830be38cbb52c7b8aa95f1ad080341de40250d17