Don't generate layouts on non live items
Non live items in Zuul queues never run jobs. They are dependency tree
markers to ensure that we contruct the correct final git states that do
run jobs. Previously we were generating layouts (necessary for running
jobs) on all of the non live items in queues which generated (n^2+n)/2
layouts when we only need a single layout for the terminal item.

Fix this by only generating layouts on live items. This introduces a
slightly new process for getting the correct layout on a child item.
The live child item can:
  Use pipeline layout when itself and items ahead do not modify layout
  Use its parent layout if parent item is live
  Generate its own layout if parent item is not live

The result is some cases where zuul would treat the parent items as
having syntax errors as a merge error in the child item now become
syntax errors in the child item. This is due to waiting to generate that
layout in the child rather than step by step.

The test that checked the old behavior described above has been updated
to reflect the new behavior introduced by this change.

Change-Id: I0d650eef9f7d1d0af46afcbabd76a1f2baf74efa