Don't validate references for excluded config items
Because of the way the exclude options work in tenant project configs,
we will still parse the job objects, even if we plan to discard them
later. This leads to failed config loading if we validate references
of config options we would exclude anyway.

This can be fixed by moving the exclude filter into parseConfig where
the parsed config (which is tenant specific) is generated.

Co-Authored-By: Tobias Henkel <tobias.henkel@bmw.de>
Change-Id: I78b68c5a3c8d403f0837d154a4ac47d11876bae5