Gracefully handle corrupt local git repositories
After an unclean shutdown of the merger or executor we can leave a git
repo in an currupt state. This could happen if a new clone is not yet
written to disk and the host crashes. After that the merger cannot
recover automatically from this situation and all future operations on
this repo will fail without manual intervention.

This can be fixed by validating if the repo can be accessed. If that
fails the repo gets deleted and cloned freshly.

Change-Id: Id8644fde6762cddda64f7f320ca760e142a55fd1