dont wait infinitely for the connection to zuul_console server
If security group or firewall on the target node silently drop
connection to zuul_console daemon - do not wait infinitely and instead
show a message.

After spending days figuring out the reason for zuul_console/stream not
working properly and figuring out that the default group of the node
does not allow 19885 connection from zuul-master (in default and zuul
security group) it became clear, that some extra information is required
back to user. I am open for making it timeout configurable, but 5 sec
should be enough initially.

Change-Id: I4e395b650e83a78d66aa362a420ef0d836c80d35