Fix race in test_reconfigure_window_fixed
By not waiting until settled after the second reconfiguration,
we might get to the release() call (which releases all jobs) before
the executor server had processed the stop jobs from the
reconfiguration event.  This caused the jobs to have the 'success'
result instead of 'aborted'.  In fact, that was the usual case,
which is why we were checking for 'success' on the restarted jobs.
Only when the system was slow enough do we see what it was really
doing -- aborting the jobs.

Also, remove the release('job1') call so that both of change B's
jobs are aborted, instead of just one.

Finally, I've added an extra wait until settled after the first
reconfiguration event.  This should not be necessary, but it
seems like good practice.

Change-Id: I87e926e009e031f8866aeab87411e3487e06f978