Enable HTML5 mode and remove hash prefix for status
The ability to provide links with filters has been added in a
previous commit however due to conflicts with Angular, it would
add an extra '#!/' before that.

This patch sets the hash prefix to nothing which drops the
exclamation mark and then enables HTML5 mode which drops the
forward slash.

Once Angular is used throughout zuul-web, we'll be able to do
this in a much cleaner way, but for now this is the best
workaround to make things work with the combination of JQuery
and AngularJS.

Change-Id: I97117db8e4b5f573b3d60b94bc2d6f2be544c086