Handle multiple stats in assertReportedStats
With a statsd pipeline you batch stats for a single send to avoid too
much packet overhead.  The stats are sent in a single transaction '\n'
delineated.  Although not currently used in Zuul, it's worth handling
this to avoid the next person having to debug this again (I copied
this general statsd testing idea into a project where it is being
used).

Change-Id: Id4f6f5a6cd66581a81299ed5c67a5c49c95c9b52