Cleanup log streaming error handling
Previously we didn't reraise, continue or return after having timeout
exceptions connecting to the job log streaming executor. This would
result in subsequent exceptions trying to use the socket object that
doesn't exist. Clean this up by returning when handling this error this
will nicely shutdown the streaming thread.

We also log this state to the job's log so that end users and ops are
aware of the problem.

Change-Id: I7fcfe68197d8afdb54f134bdf0e0e8053ce4faf2