Disable rewriting links in status page with HTML5 mode
The patch to fix the ability to do filtering based on the fragment
identifier enabled HTML5 mode, however, this meant that Angular
would takeover control all the URLs and assume they were all
going to Angular componeents.

In our case, they are not, therefore, we disable the link rewriting
feature of HTML5 mode to resolve this issue.

Change-Id: I25026d8e6a424a0816d5c7863c92486172ed0b21