Accumulate errors in context managers
Rather than adding extra exception handlers to accumulate errors,
we can just pass in the accumulator to the configuration error
context manager (which already has the SourceContext information
for the error, as well as the message) and let it handle the
accumulation.

This is a POC that only changes one instance; the rest of the
exception handlers added for the error accumulators should be
able to follow this pattern as well.

Change-Id: Id80c407544a2ee9dd8b39335aba3252b4ba7350d