Also cache empty unparsed config
We currently don't cache empty unparsed config. This has the unwanted
effect that all repos/branches not containing zuul configuration are
re-queried for their config even if we don't have to. In tenants with
many such projects tenant reconfiguration is quite
expensive. Especially the GitHub Enterprise workflow is often branch &
pull. This leads to a tenant configuration on each new PR.

Change-Id: I46ffd85866a6453dbd767e81a0e2e98609929489