Log github delivery ids properly
Previously we were doing a get for 'X-GitHub-Delivery' on a python dict
which didn't have this key. The key in the dict is 'x-github-delivery'.
Because of this we were logging None as the delivery id. Fix this by
updating the key used which will give us logging data that is helpful
for debugging.

Side note. The header as sent by github to the zuul web server appears
to be 'X-GitHub-Delivery' with that case. But somewhere along the line
things get normalized to lower case. This is true for the other headers
we lookat in the events as well.

Change-Id: I8acbc910e7e50aa53c1ade0fa2ce6f2afed053d7