Log more information about what events trigger a reconfig
A tenant reconfiguration might be caused by a branch creation,
deletion, or change to files.  Cause TriggerEvents to repr themselves
with flags indicating which of those are true.  Have the
GithubTriggerEvent also repr itself with the associated delivery id
for easier tracing.  Finally, have the scheduler log these events
when performing a tenant reconfiguration so we can see all the info.

Change-Id: I2cfc7615dfb3533aeb9d42b009cc67ce88a5ee48