Fix missing dequeue on synchronized pull request
When updating a pull request Github can automatically dismiss a
review. In this case Github also sends a pull_request_review
event. This event contains the old head of the pull request which
leads to zuul updating its internal state of the old pull
request. This also updates the updated_at timestamp. Unlike Gerrit
there is no notion of a by time ordering of pull request 'versions' so
we currently rely on the updated_at timestamp to check if a pull
request is newer than another one. So in cases like mentioned above
that fails because the old and new pull request object finally get the
same timestamp which prevents dequeueing of the old pull request.

As the updated_at timestamp is a moving target that is not bound to
the pull request 'version' we can solve that by just not updating the
timestamp if the pull request is updated in the cache. This way the
old pull request object retains its old timestamp and the update check
works.

Change-Id: Ie194ce6571942b5c175709daa3517f68c7c48379