Consolidate navigation functions
We had getBaseHref and getBaseHrefFromPath in two places, but also it
looked like they needed to do exactly the same thing.

We need getApiBaseHref to be a function so we can all it in the app.module
provider. We also want it to return a path, not a URL. (Ignore the
'href' in the name)

We need to be able to get the base URL to use for API calls - and we
need, for now, while we have links being made outside of angular, to be
able to get something similar to getApiBaseHref but that returns a url
and not a path.

Change-Id: I7457a7b160a312ea3eec74b8a4f3380a1f95d8e1