Fix logging of clone url
In the executor log we currently log the remote url of the project
instead of the real url from where we clone a copy which can be the
local cache. This leads to misinterpretation of the log which indicates
that on each job a clean clone is done. This is not the case so fix the
logging here.

Change-Id: I694c04c73fc39479d68fd3eb7268ddf8cdde4ba2