Reduce requests to github when listing project branches
In getProjectBranches in the GitHub driver first gets the repository
object and then iterates over the branches. Creating the repository
object requires a request to GitHub to get the repository
metadata. However we don't need that data and can skip this request if
we do the branch listing request by our own. This will save about half
of the requests required to get the project branches during a
reconfiguration.

Change-Id: Icd96831d467c1562cdec9d4220154c19a2e402a6