Fix source context in split config
If a repo has multiple config files, we would previously share the
source context between all of those files.  However, one of the
attributes of a source context is the path from which the config
file was loaded, so debugging statements would show that all of the
config objects loaded from that repo came from the last filename
Zuul read.

Instead, make a copy of the source context for each path loaded.

Change-Id: Ib22da429b484a99c86c2e620d97f655710460df4