Require tenant in Pipeline constructor
Pipelines are closely associated with their tenants, so that
pipeline managers may find the currently active layout.  To ensure
that a Pipeline is always associated with exactly one tenant,
require the tenant when constructing the pipeline, and raise an
exception if we try to add the pipeline to a layout for a different
tenant object.

Change-Id: I9fafa0133d57c77ee172245d0897e375dcd0de15