Add command socket handler for full reconfiguration
In future we want to support different types of reconfigurations so
only relying on signals won't scale. Thus we should make the full
reconfiguration available via the command socket which will be
extensible in the future. A later change will add a reconfiguration
without clearing the cache to be able to quuickly add or remove
projects from the tenant configuration without having too much impact
into the system.

Change-Id: I9748ecbcffa8c9b65f98d8768735bdf00e78cf25