Fix broken templates for projects in dependent pipelines
A project referencing a template which did not exist could cause
an exception during reconfiguration causing it to fail.

If a project references a template which does not exist, and Zuul
then attempts to add that project to a dependent pipeline, or
build a job graph for it, an uncaught exception will be thrown
(including during the final phase of reconfiguration).

It is safest to treat it as if the project were unconfigured until
the problem is corrected, so this change causes getAllProjectConfigs
to return the empty list if there is such an error.

Change-Id: I36c011a7c17646d20c816e44a89f075415217bfb