Reduce number of reconfigurations on branch delete
If a branch is deleted that didn't contain configuration files or was
disregarded due to its protection state, we do not need to trigger a
tenant reconfiguration. This patch checks whether there is a cached
configuration and suppresses reconfiguration on branch deletion if there
is no configuration to be removed for this branch.

Change-Id: I34f5d3631d8112c792a72146268dad85813ceb4f