Emit a log line containing version info around startup
It's useful to know which version of zuul is running, and sometimes for
forensics which version was running previously. There isn't a great
hook point only in run where we have logging configured. However, we
setup logging at least once (and sometimes more than once) for each
daemon app. Take advantage of that and emit a log line when we do.

Change-Id: I7f5884ca59567d24b5ffd7a2f20bcd3b332752e4