doc: Move zuul variable references to a section
The "variables" section discusses the 4 sources of variables
(site-wide, secrets, job and parent job) but then the zuul.* variable
reference has ended up in the middle of them.

Extract the the zuul variable reference and move it up a level, after
the discussion of the other variables.  Enumerate the list of
variables and add a few links while we're here.

Change-Id: I127c5c3730e950e11695cb6e1b660537b4bc6107