Point Suse users are zookeeper releases page
There do not appear to be official apache zookeeper packages for
opensuse, but there are official tarballs. Rather than point to one of
many zookeeper release mirrors point users to the main releases page and
intruct them to follow directions there to find the best place to
download a zookeeper tarball.

Change-Id: Ia81941a3aa250ee7254e68ae171fdafc72c20e03