Fix wrong matched project template
When having the two branches stable and stable-foo with distinct
project pipelines that pull in different project-templates currently
both project-templates match. The project-templates in this case get
implicit branch matchers. However the implied branch matcher currently
only does a partial match and so a change to stable-foo matches the
project-templates of both stable and stable-foo.

Fix this by changing the implied branch matchers to do full matching.

Change-Id: I569db50f05c1b53cb1fb26e84ea0af377e3df73c